Navigation, size contract, ZIP structure, XML ceilings — and a wrong stat-cache offset found by the new gate
Stages 9, 10, 11 and 13. Two new gates, and one of them immediately found a real bug in this tree.

STAGE 9 — navigation. Every render reply (including a CACHE HIT) now carries `sectionComplete`, and the
page reads it as `=== true`, so an unknown value fails CLOSED. Before, the worker fell back to
`complete: true` when no estimate existed — i.e. before the first progress message — so the one direction
that always "worked" (a fast Next) could cross into the next chapter while the current spine was still
building. Backward crossing now asks for `progressive: false`: "last page of the previous chapter" is only
knowable once that spine is FULLY laid out, and the progressive path built one page and let the over-range
clamp resolve to page 0, so going back a chapter landed on its FIRST page.

STAGE 10 — the file-size contract. The module is memory32 and its file I/O is `int`
(HalFile::available/read), so a book above INT_MAX is representable in neither; validateBookInput() now
rejects it at the boundary, before arrayBuffer(), _ko_epub_alloc, checkedMalloc or the external mount see
the Blob. Number.isSafeInteger() was never that check. HalFile::available() clamps instead of truncating
into a negative "not available", read() clamps the transferred count the same way, and write() guards
`pos_ + count` against wrapping into a small resize.

STAGE 11 — ZIP structure. Multi-disk archives are rejected outright (every offset in them is relative to a
disk we do not have). The central directory must now END BEFORE THE EOCD RECORD, not merely before
physical EOF. getDataOffset() validates the DATA RANGE it returns, not just the 30-byte header it computed
it from — a hostile filename/extra length used to push the offset past EOF and the caller inflated from
there. Cumulative spine sizes are checked before they are accumulated, and SizeTarget::index is uint32_t:
16 bits wrapped above 65535 spines, and a wrapped index writes the size into a DIFFERENT spine's slot.

STAGE 13 — XML ceilings (container.xml 1 MiB, OPF 16 MiB, NAV/NCX 16 MiB, cover wrapper 4 MiB), each
rejected BEFORE the parser is constructed, matching the CSS ceiling that already existed; and rename() now
moves EVERY alias of the same Blob, the way remove() already did — the mount functions install a basename
alias for every nested mount, so moving only the named key left the old name holding the file.

THE BUG THE NEW GATE FOUND. scripts/verify/zip_hardening_gate.py mutates a real EPUB one specific way per
case and requires the engine to REJECT it. Writing it, my own mutant crashed on garbage, which turned out
to be the engine's fault, not the test's: LoadAllFileStatSlims() read the local-header offset from central
record byte 44. It is at byte 42 — measured, offset 42 of a real record points at PK\x03\x04 for every
entry while byte 44 is noise. The stat cache has been filling with wrong offsets the whole time, and no
fixture noticed because their reads go through the lazy per-name path. Fixed, and the mutation now targets
the OPF's own local header rather than "the first one in the file", because a header the reader never
resolves proves nothing.

GATES. zip_hardening_gate.py: 6/6 — the healthy book loads and every hostile archive (multi-disk, CD past
the EOCD, corrupt CD record, bogus local-header lengths, 2 GiB OPF) is rejected with no container
produced. State integrity (from the previous commit): 29/29. image_once, incremental, pool, mono_planes,
external, spine_href, defensive, package_consistency and wasm smoke all PASS. Byte-neutrality against the
recorded reference: demo-images 2-bit 03b5368992e27011c3401c2b and 1-bit abe2a5feb999371015aa74e4 both
MATCH.

ORACLE. The pin correctly refused to certify: I edited two vendored engine files (Epub.cpp, BookMetadataCache.cpp)
on purpose, so they can no longer be byte-identical. Re-derived the pin, which moves them to function-level
comparison — where every function in LAYOUT_CRITICAL is still required to be identical, and is. That
requirement is what makes "the divergence did not touch typography" a check instead of a hope. Pin check
now PASSes: 152 identical, 15 divergent at function level, 1 header absent.

+++ vendor-lib/ZipFile/ZipFile.h
@@ -26,7 +26,8 @@ class ZipFile {
   struct SizeTarget {
     uint64_t hash;   // FNV-1a 64-bit hash of normalized path
     uint16_t len;    // Length of path for collision reduction
-    uint16_t index;  // Caller's index (e.g. spine index)
+    uint32_t index;  // Caller's index (e.g. spine index). 16 bits wrapped above 65535 spines, and a
+                     // wrapped index writes the size into a DIFFERENT spine's slot.
   };
 
   // FNV-1a 64-bit hash computed from char buffer (no std::string allocation)
